wxGTK compiles (and seems to work) again after wxMenu changes
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@4300 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/menucmn.cpp
@@ -310,7 +310,7 @@ wxMenuItem *wxMenuBase::FindChildItem(int id, size_t *ppos) const
 
     if ( ppos )
     {
-        *ppos = item ? pos : wxNOT_FOUND;
+        *ppos = item ? pos : (size_t)wxNOT_FOUND;
     }
 
     return item;
